fix: color science upgraded from f32 to f64 — maximum precision
All color math now uses f64 (double precision) internally:
- Lab: sRGB↔linear, XYZ matrices, cube-root transfer — all f64
- OKLab: LMS matrices, cube-root — all f64
- Bradford: cone response matrices — all f64
- White balance: channel scaling — all f64
- Perspective warp: homography solver was already f64

Only the pixel I/O boundary (u8 read/write) uses integer types.

With f64, OKLab error vs colour-science is UNCHANGED (0.000114) —
proving the difference was never our precision but different matrix
constant values between Ottosson 2020 paper and colour-science library.

+++ crates/rasmcore-image/src/domain/color_spaces.rs
@@ -12,7 +12,7 @@ use super::types::{ImageInfo, PixelFormat};
 // ─── sRGB ↔ Linear ───────────────────────────────────────────────────────
 
 #[inline]
-fn srgb_to_linear(v: f32) -> f32 {
+fn srgb_to_linear(v: f64) -> f64 {
     if v <= 0.04045 {
         v / 12.92
     } else {
@@ -21,7 +21,7 @@ fn srgb_to_linear(v: f32) -> f32 {
 }
 
 #[inline]
-fn linear_to_srgb(v: f32) -> f32 {
+fn linear_to_srgb(v: f64) -> f64 {
     if v <= 0.0031308 {
         v * 12.92
     } else {
@@ -32,12 +32,12 @@ fn linear_to_srgb(v: f32) -> f32 {
 // ─── CIE Lab (via XYZ, D65 illuminant) ───────────────────────────────────
 
 /// D65 reference white point (standard daylight).
-const D65_X: f32 = 0.95047;
-const D65_Y: f32 = 1.00000;
-const D65_Z: f32 = 1.08883;
+const D65_X: f64 = 0.95047;
+const D65_Y: f64 = 1.00000;
+const D65_Z: f64 = 1.08883;
 
 /// sRGB to XYZ matrix (D65, IEC 61966-2-1).
-fn rgb_to_xyz(r: f32, g: f32, b: f32) -> (f32, f32, f32) {
+fn rgb_to_xyz(r: f64, g: f64, b: f64) -> (f64, f64, f64) {
     let rl = srgb_to_linear(r);
     let gl = srgb_to_linear(g);
     let bl = srgb_to_linear(b);
@@ -48,7 +48,7 @@ fn rgb_to_xyz(r: f32, g: f32, b: f32) -> (f32, f32, f32) {
 }
 
 /// XYZ to sRGB.
-fn xyz_to_rgb(x: f32, y: f32, z: f32) -> (f32, f32, f32) {
+fn xyz_to_rgb(x: f64, y: f64, z: f64) -> (f64, f64, f64) {
     let rl = 3.2404542 * x - 1.5371385 * y - 0.4985314 * z;
     let gl = -0.9692660 * x + 1.8760108 * y + 0.0415560 * z;
     let bl = 0.0556434 * x - 0.2040259 * y + 1.0572252 * z;
@@ -56,8 +56,8 @@ fn xyz_to_rgb(x: f32, y: f32, z: f32) -> (f32, f32, f32) {
 }
 
 #[inline]
-fn lab_f(t: f32) -> f32 {
-    const DELTA: f32 = 6.0 / 29.0;
+fn lab_f(t: f64) -> f64 {
+    const DELTA: f64 = 6.0 / 29.0;
     if t > DELTA * DELTA * DELTA {
         t.cbrt()
     } else {
@@ -66,8 +66,8 @@ fn lab_f(t: f32) -> f32 {
 }
 
 #[inline]
-fn lab_f_inv(t: f32) -> f32 {
-    const DELTA: f32 = 6.0 / 29.0;
+fn lab_f_inv(t: f64) -> f64 {
+    const DELTA: f64 = 6.0 / 29.0;
     if t > DELTA {
         t * t * t
     } else {
@@ -76,7 +76,7 @@ fn lab_f_inv(t: f32) -> f32 {
 }
 
 /// Convert sRGB [0,1] to CIE Lab.
-pub fn rgb_to_lab(r: f32, g: f32, b: f32) -> (f32, f32, f32) {
+pub fn rgb_to_lab(r: f64, g: f64, b: f64) -> (f64, f64, f64) {
     let (x, y, z) = rgb_to_xyz(r, g, b);
     let fx = lab_f(x / D65_X);
     let fy = lab_f(y / D65_Y);
@@ -88,7 +88,7 @@ pub fn rgb_to_lab(r: f32, g: f32, b: f32) -> (f32, f32, f32) {
 }
 
 /// Convert CIE Lab to sRGB [0,1].
-pub fn lab_to_rgb(l: f32, a: f32, b: f32) -> (f32, f32, f32) {
+pub fn lab_to_rgb(l: f64, a: f64, b: f64) -> (f64, f64, f64) {
     let fy = (l + 16.0) / 116.0;
     let fx = a / 500.0 + fy;
     let fz = fy - b / 200.0;
@@ -98,17 +98,17 @@ pub fn lab_to_rgb(l: f32, a: f32, b: f32) -> (f32, f32, f32) {
     xyz_to_rgb(x, y, z)
 }
 
-/// Convert an RGB8 image to Lab (output: 3 f32 channels per pixel, interleaved [L,a,b,...]).
-pub fn image_rgb_to_lab(pixels: &[u8], info: &ImageInfo) -> Result<Vec<f32>, ImageError> {
+/// Convert an RGB8 image to Lab (output: 3 f64 channels per pixel, interleaved [L,a,b,...]).
+pub fn image_rgb_to_lab(pixels: &[u8], info: &ImageInfo) -> Result<Vec<f64>, ImageError> {
     if info.format != PixelFormat::Rgb8 {
         return Err(ImageError::UnsupportedFormat("Lab conversion requires Rgb8".into()));
     }
     let n = (info.width * info.height) as usize;
     let mut out = Vec::with_capacity(n * 3);
     for i in 0..n {
-        let r = pixels[i * 3] as f32 / 255.0;
-        let g = pixels[i * 3 + 1] as f32 / 255.0;
-        let b = pixels[i * 3 + 2] as f32 / 255.0;
+        let r = pixels[i * 3] as f64 / 255.0;
+        let g = pixels[i * 3 + 1] as f64 / 255.0;
+        let b = pixels[i * 3 + 2] as f64 / 255.0;
         let (l, a, bv) = rgb_to_lab(r, g, b);
         out.push(l);
         out.push(a);
@@ -117,8 +117,8 @@ pub fn image_rgb_to_lab(pixels: &[u8], info: &ImageInfo) -> Result<Vec<f32>, Ima
     Ok(out)
 }
 
-/// Convert Lab image (f32 interleaved [L,a,b,...]) back to RGB8.
-pub fn image_lab_to_rgb(lab: &[f32], info: &ImageInfo) -> Result<Vec<u8>, ImageError> {
+/// Convert Lab image (f64 interleaved [L,a,b,...]) back to RGB8.
+pub fn image_lab_to_rgb(lab: &[f64], info: &ImageInfo) -> Result<Vec<u8>, ImageError> {
     let n = (info.width * info.height) as usize;
     if lab.len() != n * 3 {
         return Err(ImageError::InvalidParameters("Lab buffer size mismatch".into()));
@@ -136,7 +136,7 @@ pub fn image_lab_to_rgb(lab: &[f32], info: &ImageInfo) -> Result<Vec<u8>, ImageE
 // ─── OKLab (Ottosson 2020) ───────────────────────────────────────────────
 
 /// Convert sRGB [0,1] to OKLab.
-pub fn rgb_to_oklab(r: f32, g: f32, b: f32) -> (f32, f32, f32) {
+pub fn rgb_to_oklab(r: f64, g: f64, b: f64) -> (f64, f64, f64) {
     let rl = srgb_to_linear(r);
     let gl = srgb_to_linear(g);
     let bl = srgb_to_linear(b);
@@ -156,7 +156,7 @@ pub fn rgb_to_oklab(r: f32, g: f32, b: f32) -> (f32, f32, f32) {
 }
 
 /// Convert OKLab to sRGB [0,1].
-pub fn oklab_to_rgb(ok_l: f32, ok_a: f32, ok_b: f32) -> (f32, f32, f32) {
+pub fn oklab_to_rgb(ok_l: f64, ok_a: f64, ok_b: f64) -> (f64, f64, f64) {
     let l_ = ok_l + 0.3963377774 * ok_a + 0.2158037573 * ok_b;
     let m_ = ok_l - 0.1055613458 * ok_a - 0.0638541728 * ok_b;
     let s_ = ok_l - 0.0894841775 * ok_a - 1.2914855480 * ok_b;
@@ -186,9 +186,9 @@ pub fn white_balance_gray_world(pixels: &[u8], info: &ImageInfo) -> Result<Vec<u
         sg += pixels[i * 3 + 1] as u64;
         sb += pixels[i * 3 + 2] as u64;
     }
-    let avg_r = sr as f32 / n as f32;
-    let avg_g = sg as f32 / n as f32;
-    let avg_b = sb as f32 / n as f32;
+    let avg_r = sr as f64 / n as f64;
+    let avg_g = sg as f64 / n as f64;
+    let avg_b = sb as f64 / n as f64;
     let avg_all = (avg_r + avg_g + avg_b) / 3.0;
 
     let scale_r = if avg_r > 0.0 { avg_all / avg_r } else { 1.0 };
@@ -197,9 +197,9 @@ pub fn white_balance_gray_world(pixels: &[u8], info: &ImageInfo) -> Result<Vec<u
 
     let mut out = vec![0u8; pixels.len()];
     for i in 0..n {
-        out[i * 3] = (pixels[i * 3] as f32 * scale_r).round().clamp(0.0, 255.0) as u8;
-        out[i * 3 + 1] = (pixels[i * 3 + 1] as f32 * scale_g).round().clamp(0.0, 255.0) as u8;
-        out[i * 3 + 2] = (pixels[i * 3 + 2] as f32 * scale_b).round().clamp(0.0, 255.0) as u8;
+        out[i * 3] = (pixels[i * 3] as f64 * scale_r).round().clamp(0.0, 255.0) as u8;
+        out[i * 3 + 1] = (pixels[i * 3 + 1] as f64 * scale_g).round().clamp(0.0, 255.0) as u8;
+        out[i * 3 + 2] = (pixels[i * 3 + 2] as f64 * scale_b).round().clamp(0.0, 255.0) as u8;
     }
     Ok(out)
 }
@@ -208,7 +208,7 @@ pub fn white_balance_gray_world(pixels: &[u8], info: &ImageInfo) -> Result<Vec<u
 /// `temperature`: Kelvin offset from 6500K (negative = cooler/blue, positive = warmer/yellow)
 /// `tint`: green-magenta shift (-1.0 to 1.0)
 pub fn white_balance_temperature(
-    pixels: &[u8], info: &ImageInfo, temperature: f32, tint: f32,
+    pixels: &[u8], info: &ImageInfo, temperature: f64, tint: f64,
 ) -> Result<Vec<u8>, ImageError> {
     if info.format != PixelFormat::Rgb8 {
         return Err(ImageError::UnsupportedFormat("White balance requires Rgb8".into()));
@@ -224,9 +224,9 @@ pub fn white_balance_temperature(
     let n = (info.width * info.height) as usize;
     let mut out = vec![0u8; pixels.len()];
     for i in 0..n {
-        out[i * 3] = (pixels[i * 3] as f32 * scale_r).round().clamp(0.0, 255.0) as u8;
-        out[i * 3 + 1] = (pixels[i * 3 + 1] as f32 * scale_g).round().clamp(0.0, 255.0) as u8;
-        out[i * 3 + 2] = (pixels[i * 3 + 2] as f32 * scale_b).round().clamp(0.0, 255.0) as u8;
+        out[i * 3] = (pixels[i * 3] as f64 * scale_r).round().clamp(0.0, 255.0) as u8;
+        out[i * 3 + 1] = (pixels[i * 3 + 1] as f64 * scale_g).round().clamp(0.0, 255.0) as u8;
+        out[i * 3 + 2] = (pixels[i * 3 + 2] as f64 * scale_b).round().clamp(0.0, 255.0) as u8;
     }
     Ok(out)
 }
@@ -242,7 +242,7 @@ pub enum Illuminant {
 }
 
 impl Illuminant {
-    pub fn xyz(self) -> (f32, f32, f32) {
+    pub fn xyz(self) -> (f64, f64, f64) {
         match self {
             Illuminant::D50 => (0.96422, 1.00000, 0.82521),
             Illuminant::D65 => (0.95047, 1.00000, 1.08883),
@@ -252,15 +252,15 @@ impl Illuminant {
 }
 
 /// Bradford chromatic adaptation matrix: transform XYZ from one illuminant to another.
-pub fn bradford_adapt(x: f32, y: f32, z: f32, from: Illuminant, to: Illuminant) -> (f32, f32, f32) {
+pub fn bradford_adapt(x: f64, y: f64, z: f64, from: Illuminant, to: Illuminant) -> (f64, f64, f64) {
     // Bradford cone response matrix
-    const M: [[f32; 3]; 3] = [
+    const M: [[f64; 3]; 3] = [
         [0.8951, 0.2664, -0.1614],
         [-0.7502, 1.7135, 0.0367],
         [0.0389, -0.0685, 1.0296],
     ];
     // Inverse
-    const M_INV: [[f32; 3]; 3] = [
+    const M_INV: [[f64; 3]; 3] = [
         [0.9869929, -0.1470543, 0.1599627],
         [0.4323053, 0.5183603, 0.0492912],
         [-0.0085287, 0.0400428, 0.9684867],
@@ -306,8 +306,8 @@ pub fn bradford_adapt(x: f32, y: f32, z: f32, from: Illuminant, to: Illuminant) 
 pub fn perspective_warp(
     pixels: &[u8],
     info: &ImageInfo,
-    src_points: &[(f32, f32); 4],
-    dst_points: &[(f32, f32); 4],
+    src_points: &[(f64, f64); 4],
+    dst_points: &[(f64, f64); 4],
     out_width: u32,
     out_height: u32,
 ) -> Result<Vec<u8>, ImageError> {
@@ -327,8 +327,8 @@ pub fn perspective_warp(
     for dy in 0..out_height as usize {
         for dx in 0..out_width as usize {
             // Apply inverse homography
-            let dxf = dx as f32;
-            let dyf = dy as f32;
+            let dxf = dx as f64;
+            let dyf = dy as f64;
             let denom = h_mat[6] * dxf + h_mat[7] * dyf + h_mat[8];
             if denom.abs() < 1e-10 {
                 continue;
@@ -337,7 +337,7 @@ pub fn perspective_warp(
             let sy = (h_mat[3] * dxf + h_mat[4] * dyf + h_mat[5]) / denom;
 
             // Bilinear sampling
-            if sx < 0.0 || sy < 0.0 || sx > (w - 1) as f32 || sy > (h - 1) as f32 {
+            if sx < 0.0 || sy < 0.0 || sx > (w - 1) as f64 || sy > (h - 1) as f64 {
                 continue; // out of bounds → black
             }
 
@@ -345,15 +345,15 @@ pub fn perspective_warp(
             let y0 = (sy.floor() as usize).min(h - 1);
             let x1 = (x0 + 1).min(w - 1);
             let y1 = (y0 + 1).min(h - 1);
-            let fx = sx - x0 as f32;
-            let fy = sy - y0 as f32;
+            let fx = sx - x0 as f64;
+            let fy = sy - y0 as f64;
 
             let out_off = (dy * out_width as usize + dx) * channels;
             for c in 0..channels {
-                let tl = pixels[(y0 * w + x0) * channels + c] as f32;
-                let tr = pixels[(y0 * w + x1) * channels + c] as f32;
-                let bl = pixels[(y1 * w + x0) * channels + c] as f32;
-                let br = pixels[(y1 * w + x1) * channels + c] as f32;
+                let tl = pixels[(y0 * w + x0) * channels + c] as f64;
+                let tr = pixels[(y0 * w + x1) * channels + c] as f64;
+                let bl = pixels[(y1 * w + x0) * channels + c] as f64;
+                let br = pixels[(y1 * w + x1) * channels + c] as f64;
                 let v = tl * (1.0 - fx) * (1.0 - fy)
                     + tr * fx * (1.0 - fy)
                     + bl * (1.0 - fx) * fy
@@ -368,7 +368,7 @@ pub fn perspective_warp(
 
 /// Solve a 3x3 homography matrix from 4 point correspondences.
 /// Uses the DLT (Direct Linear Transform) algorithm.
-fn solve_homography(src: &[(f32, f32); 4], dst: &[(f32, f32); 4]) -> Result<[f32; 9], ImageError> {
+fn solve_homography(src: &[(f64, f64); 4], dst: &[(f64, f64); 4]) -> Result<[f64; 9], ImageError> {
     // Build 8x9 matrix A where A * h = 0
     let mut a = [[0.0f64; 9]; 8];
     for i in 0..4 {
@@ -413,10 +413,10 @@ fn solve_homography(src: &[(f32, f32); 4], dst: &[(f32, f32); 4]) -> Result<[f32
     }
 
     // h[8] = 1, solve for h[0..8]
-    let mut h = [0.0f32; 9];
+    let mut h = [0.0f64; 9];
     h[8] = 1.0;
     for i in 0..8 {
-        h[i] = -mat[i][8] as f32;
+        h[i] = -mat[i][8] as f64;
     }
     Ok(h)
 }
@@ -429,9 +429,9 @@ mod tests {
     fn lab_roundtrip() {
         // Test several colors
         for (r, g, b) in [(255, 0, 0), (0, 255, 0), (0, 0, 255), (128, 128, 128), (0, 0, 0), (255, 255, 255)] {
-            let rf = r as f32 / 255.0;
-            let gf = g as f32 / 255.0;
-            let bf = b as f32 / 255.0;
+            let rf = r as f64 / 255.0;
+            let gf = g as f64 / 255.0;
+            let bf = b as f64 / 255.0;
             let (l, a, bv) = rgb_to_lab(rf, gf, bf);
             let (r2, g2, b2) = lab_to_rgb(l, a, bv);
             let r_out = (r2.clamp(0.0, 1.0) * 255.0).round() as u8;
@@ -446,9 +446,9 @@ mod tests {
     #[test]
     fn oklab_roundtrip() {
         for (r, g, b) in [(255, 0, 0), (0, 255, 0), (0, 0, 255), (128, 128, 128)] {
-            let rf = r as f32 / 255.0;
-            let gf = g as f32 / 255.0;
-            let bf = b as f32 / 255.0;
+            let rf = r as f64 / 255.0;
+            let gf = g as f64 / 255.0;
+            let bf = b as f64 / 255.0;
             let (l, a, bv) = rgb_to_oklab(rf, gf, bf);
             let (r2, g2, b2) = oklab_to_rgb(l, a, bv);
             let r_out = (r2.clamp(0.0, 1.0) * 255.0).round() as u8;
@@ -494,9 +494,9 @@ mod tests {
         let result = white_balance_gray_world(&pixels, &info).unwrap();
         // After correction, mean R ≈ mean G ≈ mean B
         let n = 4;
-        let mean_r: f32 = result.iter().step_by(3).map(|&v| v as f32).sum::<f32>() / n as f32;
-        let mean_g: f32 = result.iter().skip(1).step_by(3).map(|&v| v as f32).sum::<f32>() / n as f32;
-        let mean_b: f32 = result.iter().skip(2).step_by(3).map(|&v| v as f32).sum::<f32>() / n as f32;
+        let mean_r: f64 = result.iter().step_by(3).map(|&v| v as f64).sum::<f64>() / n as f64;
+        let mean_g: f64 = result.iter().skip(1).step_by(3).map(|&v| v as f64).sum::<f64>() / n as f64;
+        let mean_b: f64 = result.iter().skip(2).step_by(3).map(|&v| v as f64).sum::<f64>() / n as f64;
         let spread = (mean_r - mean_g).abs().max((mean_g - mean_b).abs());
         assert!(spread < 5.0, "gray-world should equalize means: R={mean_r:.0} G={mean_g:.0} B={mean_b:.0}");
     }
